requirements.txt
Pillow==8.4.0
certifi==2020.6.20
chardet==4.0.0
feedparser==6.0.8
httplib2==0.20.2
idna==2.10
mutagen==1.45.1
pycairo==1.20.1
pycurl==7.44.1
PyGObject==3.42.0
pyparsing==2.4.7
PySimpleSOAP==1.16.2
python-apt==2.3.0+b1
python-debian==0.1.42
python-debianbts==3.2.0
quodlibet==4.4.0
reportbug==11.1.0
requests==2.25.1
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.5
